Session populates db CardCounters update

+++ FlashCards.DB/dbo/StudySession.sql
@@ -3,7 +3,7 @@
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY,
 	[Time] DATETIME NOT NULL,
 	[StackId] INT NOT NULL,
-	[Score]  INT NOT NULL,
+	[Score]  FLOAT NOT NULL,
 	[CountStudied] INT NOT NULL,
 	[CountCorrect] INT NOT NULL,
 	[CountIncorrect] INT NOT NULL,
